Pull various types out of `sync15` into a new crate
