+++ components/support/sync15-traits/README.md
@@ -1,0 +1,4 @@
+# sync15-traits
+
+Extracted types and traits from sync15. Usable for cases where depending on the
+sync15 crate is impossible (like in remerge).
