+++ components/sync15/Cargo.toml
@@ -25,3 +25,4 @@ viaduct = { path = "../viaduct" }
 interrupt = { path = "../support/interrupt" }
 error-support = { path = "../support/error" }
 sync-guid = { path = "../support/guid", features = ["rusqlite_support", "random"] }
+sync15-traits = {path = "../support/sync15-traits"}
