+++ components/sync15/src/bso_record.rs
@@ -9,9 +9,9 @@ use lazy_static::lazy_static;
 use serde::de::{Deserialize, DeserializeOwned};
 use serde::ser::Serialize;
 use serde_derive::*;
-use serde_json::{self, Map, Value as JsonValue};
-use std::convert::From;
+use serde_json::Value as JsonValue;
 use std::ops::{Deref, DerefMut};
+pub use sync15_traits::Payload;
 use sync_guid::Guid;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -162,163 +162,68 @@ impl<T> DerefMut for BsoRecord<T> {
     }
 }
 
-/// Represents the decrypted payload in a Bso. Provides a minimal layer of type safety to avoid double-encrypting.
-///
-/// Note: If we implement a full sync client in rust we may want to consider using stronger types for each record
-/// (we did this in the past as well), but for now, since everything is just going over the FFI, there's not a lot of
-/// benefit here.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Payload {
-    pub id: Guid,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "is_false")]
-    pub deleted: bool,
-
-    #[serde(flatten)]
-    pub data: Map<String, JsonValue>,
-}
-
-// TODO: Move skip_if_default from places to something shared
-#[inline]
-#[allow(clippy::trivially_copy_pass_by_ref)]
-fn is_false(b: &bool) -> bool {
-    !*b
-}
-
-impl Payload {
-    #[inline]
-    pub fn new_tombstone(id: String) -> Payload {
-        Payload {
-            id: id.into(),
-            deleted: true,
-            data: Map::new(),
-        }
-    }
-
-    #[inline]
-    pub fn new_tombstone_with_ttl(id: String, ttl: u32) -> Payload {
-        let mut result = Payload::new_tombstone(id);
-        result.data.insert("ttl".into(), ttl.into());
-        result
-    }
-
-    #[inline]
-    pub fn with_sortindex(mut self, index: i32) -> Payload {
-        self.data.insert("sortindex".into(), index.into());
-        self
-    }
-
-    #[inline]
-    pub fn id(&self) -> &str {
-        &self.id[..]
-    }
-
-    #[inline]
-    pub fn is_tombstone(&self) -> bool {
-        self.deleted
-    }
-
-    pub fn into_bso(mut self, collection: String) -> CleartextBso {
-        let id = self.id.clone();
-        let sortindex: Option<i32> = self.take_auto_field("sortindex");
-        let ttl: Option<u32> = self.take_auto_field("ttl");
-        CleartextBso {
+impl CleartextBso {
+    pub fn from_payload(mut payload: Payload, collection: impl Into<String>) -> Self {
+        let id = payload.id.clone();
+        let sortindex: Option<i32> = take_auto_field(&mut payload, "sortindex");
+        let ttl: Option<u32> = take_auto_field(&mut payload, "ttl");
+        BsoRecord {
             id,
-            collection,
+            collection: collection.into(),
             modified: 0.into(), // Doesn't matter.
             sortindex,
             ttl,
-            payload: self,
+            payload,
         }
-    }
-
-    pub fn from_json(value: JsonValue) -> error::Result<Payload> {
-        Ok(serde_json::from_value(value)?)
-    }
-
-    pub fn into_record<T>(self) -> Result<T, serde_json::Error>
-    where
-        for<'a> T: Deserialize<'a>,
-    {
-        Ok(serde_json::from_value(JsonValue::from(self))?)
-    }
-
-    pub fn from_record<T: Serialize>(v: T) -> error::Result<Payload> {
-        // TODO: This is dumb, we do to_value and then from_value. If we end up using this
-        // method a lot we should rethink... As it is it should just be for uploading
-        // meta/global or crypto/keys which is rare enough that it doesn't matter.
-        Ok(Payload::from_json(serde_json::to_value(v)?)?)
-    }
-
-    pub fn into_json_string(self) -> String {
-        serde_json::to_string(&JsonValue::from(self))
-            .expect("JSON.stringify failed, which shouldn't be possible")
-    }
-
-    /// "Auto" fields are fields like 'sortindex' (and potentially 'ttl' in
-    /// the future) which are:
-    ///
-    /// - Added to the payload automatically when deserializing if present on
-    ///   the incoming BSO.
-    /// - Removed from the payload automatically and attached to the BSO if
-    ///   present on the outgoing payload.
-    fn add_auto_field<T: Into<JsonValue>>(&mut self, name: &str, v: Option<T>) {
-        // This is a little dubious, but it seems like if we have a e.g. `sortindex` field on the payload
-        // it's going to be a bug if we use it instead of the "real" sort index.
-        if self.data.contains_key(name) {
-            log::warn!(
-                "Payload for record {} already contains 'automatic' field \"{}\"? \
-                 Overwriting with 'real' value",
-                self.id,
-                name
-            );
-        }
-
-        if let Some(value) = v {
-            self.data.insert(name.into(), value.into());
-        } else {
-            self.data.remove(name);
-        }
-    }
-
-    fn take_auto_field<V>(&mut self, name: &str) -> Option<V>
-    where
-        for<'a> V: Deserialize<'a>,
-    {
-        let v = self.data.remove(name)?;
-        match serde_json::from_value(v) {
-            Ok(v) => Some(v),
-            Err(e) => {
-                log::error!(
-                    "Automatic field {} exists on payload, but cannot be deserialized: {}",
-                    name,
-                    e
-                );
-                None
-            }
-        }
-    }
-}
-
-impl From<Payload> for JsonValue {
-    fn from(cleartext: Payload) -> Self {
-        let Payload {
-            mut data,
-            id,
-            deleted,
-        } = cleartext;
-        data.insert("id".to_string(), JsonValue::String(id.into_string()));
-        if deleted {
-            data.insert("deleted".to_string(), JsonValue::Bool(true));
-        }
-        JsonValue::Object(data)
     }
 }
 
 pub type EncryptedBso = BsoRecord<EncryptedPayload>;
 pub type CleartextBso = BsoRecord<Payload>;
 
+/// "Auto" fields are fields like 'sortindex' (and potentially 'ttl' in
+/// the future) which are:
+///
+/// - Added to the payload automatically when deserializing if present on
+///   the incoming BSO.
+/// - Removed from the payload automatically and attached to the BSO if
+///   present on the outgoing payload.
+fn add_auto_field<T: Into<JsonValue>>(p: &mut Payload, name: &str, v: Option<T>) {
+    // This is a little dubious, but it seems like if we have a e.g. `sortindex` field on the payload
+    // it's going to be a bug if we use it instead of the "real" sort index.
+    if p.data.contains_key(name) {
+        log::warn!(
+            "Payload for record {} already contains 'automatic' field \"{}\"? \
+             Overwriting with 'real' value",
+            p.id,
+            name
+        );
+    }
+
+    if let Some(value) = v {
+        p.data.insert(name.into(), value.into());
+    } else {
+        p.data.remove(name);
+    }
+}
+
+fn take_auto_field<V>(p: &mut Payload, name: &str) -> Option<V>
+where
+    for<'a> V: Deserialize<'a>,
+{
+    let v = p.data.remove(name)?;
+    match serde_json::from_value(v) {
+        Ok(v) => Some(v),
+        Err(e) => {
+            log::error!(
+                "Automatic field {} exists on payload, but cannot be deserialized: {}",
+                name,
+                e
+            );
+            None
+        }
+    }
+}
 // Contains the methods to automatically deserialize the payload to/from json.
 mod as_json {
     use serde::de::{self, Deserialize, DeserializeOwned, Deserializer};
@@ -394,8 +299,8 @@ impl EncryptedBso {
     pub fn decrypt(self, key: &KeyBundle) -> error::Result<CleartextBso> {
         let mut new_payload: Payload = self.payload.decrypt_and_parse_payload(key)?;
         // This is a slightly dodgy place to do this, but whatever.
-        new_payload.add_auto_field("sortindex", self.sortindex);
-        new_payload.add_auto_field("ttl", self.ttl);
+        add_auto_field(&mut new_payload, "sortindex", self.sortindex);
+        add_auto_field(&mut new_payload, "ttl", self.ttl);
 
         let result = self.with_payload(new_payload);
         Ok(result)
@@ -488,9 +393,11 @@ mod tests {
 
     #[test]
     fn test_roundtrip_crypt_tombstone() {
-        let orig_record = Payload::from_json(json!({ "id": "aaaaaaaaaaaa", "deleted": true, }))
-            .unwrap()
-            .into_bso("dummy".into());
+        let orig_record = CleartextBso::from_payload(
+            Payload::from_json(json!({ "id": "aaaaaaaaaaaa", "deleted": true, })).unwrap(),
+            "dummy",
+        );
+
         assert!(orig_record.is_tombstone());
 
         let keybundle = KeyBundle::new_random().unwrap();
@@ -514,9 +421,8 @@ mod tests {
     #[test]
     fn test_roundtrip_crypt_record() {
         let payload = json!({ "id": "aaaaaaaaaaaa", "age": 105, "meta": "data" });
-        let orig_record = Payload::from_json(payload.clone())
-            .unwrap()
-            .into_bso("dummy".into());
+        let orig_record =
+            CleartextBso::from_payload(Payload::from_json(payload.clone()).unwrap(), "dummy");
 
         assert!(!orig_record.is_tombstone());
 
@@ -541,9 +447,7 @@ mod tests {
     #[test]
     fn test_record_auto_fields() {
         let payload = json!({ "id": "aaaaaaaaaaaa", "age": 105, "meta": "data", "sortindex": 100, "ttl": 99 });
-        let bso = Payload::from_json(payload)
-            .unwrap()
-            .into_bso("dummy".into());
+        let bso = CleartextBso::from_payload(Payload::from_json(payload).unwrap(), "dummy");
 
         // We don't want the keys ending up in the actual record data on the server.
         assert!(!bso.payload.data.contains_key("sortindex"));
@@ -567,9 +471,7 @@ mod tests {
     #[test]
     fn test_record_bad_hmac() {
         let payload = json!({ "id": "aaaaaaaaaaaa", "age": 105, "meta": "data", "sortindex": 100, "ttl": 99 });
-        let bso = Payload::from_json(payload)
-            .unwrap()
-            .into_bso("dummy".into());
+        let bso = CleartextBso::from_payload(Payload::from_json(payload).unwrap(), "dummy");
 
         let keybundle = KeyBundle::new_random().unwrap();
         let encrypted = bso.encrypt(&keybundle).unwrap();
